Wordle updates

+++ resources/wordle-timetravel.js
@@ -121,7 +121,7 @@ window.timeMachine = {
 
     getDay: (day) => {
         let existingDay = JSON.parse(window.localStorage.getItem(`gameState_${day}`));
-        if (existingDay) {
+        if (existingDay && existingDay.solution === timeMachine.puzzles[day]) {
             window.localStorage.setItem('gameState', JSON.stringify(existingDay));
         } else if (timeMachine.puzzles && timeMachine.puzzles[day]) {
             let gameState = Object.assign(JSON.parse(window.localStorage.getItem('gameState')), {
